+++ docker/terraform-mod/modules/bucket/outputs.tf
@@ -1,0 +1,9 @@
+output "bucket_arn" {
+  value = aws_s3_bucket.main.arn
+}
+output "bucket_id" {
+  value = aws_s3_bucket.main.id
+}
+output "bucket_regional_domain_name" {
+  value = aws_s3_bucket.main.bucket_regional_domain_name
+}